requirements.txt
numpy
scipy
tensorboard
librosa
numba
pytorch-lightning
gradio
gradio_client
ffmpeg-python
onnxruntime
tqdm
funasr
cn2an
pypinyin
pyopenjtalk
g2p_en
torchaudio
sentencepiece
transformers
chardet
PyYAML
psutil
jieba_fast
jieba
LangSegment>=0.2.0
Faster_Whisper